convert to single page
